add module icon theme validator
