+++ packages/cli/lib/validators/applyValidators.js
@@ -1,9 +1,4 @@
-const ValidatorStore = require('./ValidatorStore');
-
 async function applyValidators(validators, ...args) {
-  // Clear global store
-  ValidatorStore.clear();
-
   return Promise.all(
     validators.map(async Validator => {
       const validationResult = await Validator.validate(...args);
